+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased](https://github.com/spenserblack/check-latest-rs/compare/v0.1.0...master)
+### Changed
+- Blocking requests to be enabled by `blocking` feature
+
+### Fixed
+- Documentation for `max_minor_version!` and `max_patch!` having examples for `max_version!`
 
 ## 0.1.0 2019/01/25
 ### Initial Version :tada: